+++ renku/core/interface/plan_gateway.py
@@ -26,7 +26,7 @@ from renku.domain_model.workflow.plan import AbstractPlan
 class IPlanGateway(ABC):
     """Interface for the PlanGateway."""
 
-    def get_by_id(self, id: str) -> Optional[AbstractPlan]:
+    def get_by_id(self, id: Optional[str]) -> Optional[AbstractPlan]:
         """Get a plan by id."""
         raise NotImplementedError
 
@@ -38,8 +38,8 @@ class IPlanGateway(ABC):
         """Search plans by name."""
         raise NotImplementedError
 
-    def get_newest_plans_by_names(self, with_invalidated: bool = False) -> Dict[str, AbstractPlan]:
-        """Return a list of all newest plans with their names."""
+    def get_newest_plans_by_names(self, include_deleted: bool = False) -> Dict[str, AbstractPlan]:
+        """Return a mapping of all plan names to their newest plans."""
         raise NotImplementedError
 
     def get_all_plans(self) -> List[AbstractPlan]: